Sort-of-working closures. Refs #122.
The chief problem with this implementation is that we require Function instances
to capture their arguments when the function is created. This is not very pythonic,
because we often write functions with mutual recursion, and so one of the functions
cannot be defined when the other is.

Our next step is to try to determine whether a captured variable is _stable_ at
the function's first use. This is different in the compiler vs when the function
is being passed from the interpreter to an entrypoint. In both cases, we're
interested in trying to distinguish between the case where a captured variable
can change, in which case we need to model it with some kind of mutable storage,
and the case where a captured variable can't change, in which case we can simply
pass it along with the function (which is obviously much faster). But in the
compiler case, we can perform more reasoning about the lifetime of the function
than we can in the interpreted case, because we can't see where the function
came from.